+++ CHANGELOG.md
@@ -8,3 +8,7 @@
 
 -   Changed from nx cli to list projects to using the vscode api to find the projects ourself
 -   Changed from nx cli to list targets, to reading $.targets from the project.json of the selected project
+
+## [2026-01-02]
+
+-   After you start the command from the command pallete, added a loading notification to indicate that the extension is reading your nx monorepo to discover the projects that exist.